bug report struct created and progress tracker class moved into the Other folder
