+++ ElementX/Sources/Screens/BugReport/BugReportViewModel.swift
@@ -52,11 +52,12 @@ class BugReportViewModel: BugReportViewModelType, BugReportViewModelProtocol {
         let progressTracker = ProgressTracker()
         callback?(.submitStarted(progressTracker: progressTracker))
         do {
-            let result = try await bugReportService.submitBugReport(text: context.reportText,
-                                                                    includeLogs: context.sendingLogsEnabled,
-                                                                    includeCrashLog: true,
-                                                                    githubLabels: [],
-                                                                    files: [],
+            let bugReport = BugReport(text: context.reportText,
+                                      includeLogs: context.sendingLogsEnabled,
+                                      includeCrashLog: true,
+                                      githubLabels: [],
+                                      files: [])
+            let result = try await bugReportService.submitBugReport(bugReport,
                                                                     progressTracker: progressTracker)
             MXLog.info("SubmitBugReport succeeded, result: \(result.reportUrl)")
             callback?(.submitFinished)
